fix(conversation-cache): address review feedback
